Disable automatic velocity updates on boats
Velocity on vehicles in handled entirely by client prediction. On the server, the boat's velocity is zero at all times, except for when gravity is applied.

Because of this, when gravity is applied, the boat's horizontal velocity is set to zero, which is jarring for the client and doesn't match vanilla behavior.